requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
stem==1.8.2
python-dotenv==1.0.1
aiohttp==3.9.3
asyncio==3.4.3
